code split into multiple component
