+++ src/components/MovieSearch.js
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import axios from "axios";
-import Resources from "./Resources";
-import Navbar from "./navbar";
+import Resources from "./resources";
+import Navbar from "./navbar/index";
 import Navbar2 from "./navbar/Navbar2";
 
 function MovieSearch() {
@@ -41,13 +41,11 @@ function MovieSearch() {
             },
           })
           .then((res) => {
-            // console.log(res);
             setResources(res?.data ?? {});
 
             if (res.status === 200) {
               setSuccess(true);
               setError("movie not found..");
-              //  console.log(error);
             }
           });
       } else {
@@ -63,7 +61,6 @@ function MovieSearch() {
             if (res.status === 200) {
               setSuccess(false);
               setError("Movie Not Found..");
-              //  console.log(error);
             }
           });
       }
@@ -88,14 +85,14 @@ function MovieSearch() {
 
   return (
     <div>
-        <Navbar
-          handleInputChange={(value) => setSearchText(value)}
-          value={searchText}
-          handleKeyDown={(event) => handleKeyDown(event)}
-          handleClick={() => clickHandler(searchText)}
-          handleRedioChange={(value) => setSearchMethod(value)}
-          selectedValue={searchMethod}
-        />
+      <Navbar
+        handleInputChange={(value) => setSearchText(value)}
+        value={searchText}
+        handleKeyDown={(event) => handleKeyDown(event)}
+        handleClick={() => clickHandler(searchText)}
+        handleRedioChange={(value) => setSearchMethod(value)}
+        selectedValue={searchMethod}
+      />
       <Navbar2 />
 
       <p style={{ color: "red" }}> {inputError} </p>
