+++ src/components/card/CardBottom.css
@@ -1,0 +1,98 @@
+
+  .titleReview_bar {
+    height: 31px;
+    padding-left: 20px;
+    padding-bottom: 18px;
+    box-sizing: content-box;
+  }
+  .titleReviewBarItem {
+    display: inline-block;
+  }
+  .titleReviewSubItem {
+    display: flex;
+  }
+  .titleReviewSubItem a {
+    color: #70579D;
+    text-decoration: none;
+  }
+  .score {
+    color: white;
+    font-size: 15px;
+    height: 28px;
+    text-align: center;
+    vertical-align: 55%;
+    width: 28px;
+    padding: 8px;
+    background-color: #ffcc33;
+    margin: 5px;
+  }
+  .titleReviewSubItem a {
+    color: #70579D;
+    text-decoration: none;
+  }
+  .subtext {
+    font-size: 11px;
+    color: #444;
+  }
+  .sub_item {
+    display: flex;
+    flex-direction: column;
+  }
+  .divider {
+    border-left: solid 1px #c5c5c5;
+    display: inline-block;
+    height: 32px;
+    margin: 0 5px 0 5px;
+    padding: 0;
+    vertical-align: top;
+    width: 1px;
+  }
+  .reviews {
+    display: inline-block;
+  }
+  .popularity {
+    display:inline-block;
+  }
+  .arraw-icon {
+    width: 35px;
+    height: 28px;
+  }
+  .popularity-review {
+    display: inline-block;
+  }
+  .logo {
+    display: flex;
+    border: 1px solid #cccccc;
+    padding: 10px 20px;
+  }
+  .logo img {
+    height: 22px;
+    width: 68px;
+    margin-right: 15px;
+  }
+ .logo a {
+  color: #70579D;
+  text-decoration: none;
+ }
+ .logo a:hover {
+   text-decoration: underline;
+ }
+ .award {
+   display: flex;
+   text-align: center;
+   border-top: 1px solid #cccccc;
+   background-color: #EEE3A4;
+   /* height: 2.5rem; */
+   font-family: Verdana, Arial, sans-serif;
+   padding: 10px 15px;
+ }
+ .award span {
+   font-size: 13px;
+ }
+ .award a {
+  color: #70579D;
+  text-decoration: none;
+ }
+ .award a:hover {
+  text-decoration: underline;
+ }
