+++ src/components/card/CardTop.css
@@ -1,13 +1,4 @@
-.container {
-    margin: 1rem 24rem;
-    max-width: 100%;
-  }
-  .card {
-    display: grid;
-    box-shadow: 1px 1px 5px #ffffff;
-    background-color: white;
-  }
-  .top-section {
+.top-section {
     width: 100%;
     height: 125px;
     background-color: #333;
@@ -191,104 +182,3 @@
   color: #136cb2;
   text-decoration: underline;
 }
-/* poster */
-  .my-image {
-     width: 100%; 
-    height: 300px; 
-  }
-  .card-content {
-    padding: 18px 20px;
-  }
-  .card-title {
-    margin-bottom: 0.5rem;
-    font-size: 3.2rem;
-  }
-  
-  .card-description {
-    font-size: 15px;
-  }
-  .bottom-section {
-    display:inline-block;
-    position: relative;
-  }
-  .button {
-    background: #2a84e4;
-    color: #fff;
-    border: 1px solid rgb(0,0,0,0.75);
-    border-radius: 5px;
-    font-weight: bold;
-    font-size: 1rem;
-    cursor: pointer;
-    transition: background-color 250ms;
-    margin: 0 20px 18px;
-    min-height: 3rem;
-    min-width: 360px;
-    display: flex;
-    justify-content: left;
-    align-items: center;
-    padding: 0px 10px;
-  }
- .button:hover {
-   background-color:#094d96 ;
- }
-  .titleReview_bar {
-    height: 31px;
-    padding-left: 20px;
-    padding-bottom: 18px;
-    box-sizing: content-box;
-  }
-  .titleReviewBarItem {
-    display: inline-block;
-  }
-  .titleReviewSubItem {
-    display: flex;
-  }
-  .titleReviewSubItem a {
-    color: #70579D;
-    text-decoration: none;
-  }
-  .score {
-    color: white;
-    font-size: 15px;
-    height: 28px;
-    text-align: center;
-    vertical-align: 55%;
-    width: 28px;
-    padding: 8px;
-    background-color: #ffcc33;
-    margin: 5px;
-  }
-  .titleReviewSubItem a {
-    color: #70579D;
-    text-decoration: none;
-  }
-  .subtext {
-    font-size: 11px;
-    color: #444;
-  }
-  .sub_item {
-    display: flex;
-    flex-direction: column;
-  }
-  .divider {
-    border-left: solid 1px #c5c5c5;
-    display: inline-block;
-    height: 32px;
-    margin: 0 5px 0 5px;
-    padding: 0;
-    vertical-align: top;
-    width: 1px;
-  }
-  .reviews {
-    display: inline-block;
-  }
-  .popularity {
-    display:inline-block;
-  }
-  .arraw-icon {
-    width: 35px;
-    height: 28px;
-  }
-  .popularity-review {
-    display: inline-block;
-  }
